バックエンド側リファクタリングした
