+++ containers/nginx/nginx.conf.template
@@ -8,7 +8,7 @@ upstream myapp {
 server {
   listen 80;
   # ドメインもしくはIPを指定
-  server_name $SERVER_NAME;
+  server_name localhost;
 
   access_log /var/log/nginx/access.log;
   error_log  /var/log/nginx/error.log;
@@ -19,9 +19,19 @@ server {
   client_max_body_size 100m;
   error_page 404             /404.html;
   error_page 505 502 503 504 /500.html;
-  try_files  $uri/index.html $uri @myapp;
-  keepalive_timeout 5;
-  sendfile        off;
+
+  # APIリクエストのための設定
+  location /api/ {
+    proxy_pass http://myapp;
+    proxy_set_header X-Real-IP $remote_addr;
+    proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+    proxy_set_header Host $http_host;
+  }
+
+  # その他のリクエスト用設定
+  location / {
+    try_files $uri/index.html $uri @myapp;
+  }
 
   # リバースプロキシ関連の設定
   location @myapp {
@@ -29,10 +39,8 @@ server {
     proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
     proxy_set_header Host $http_host;
     proxy_pass http://myapp;
-
-    # キャッシュを無視する設定
-    proxy_ignore_headers Cache-Control Expires;
-    proxy_no_cache 1;
-    proxy_cache_bypass 1;
   }
+
+  keepalive_timeout 5;
+  sendfile        off;
 }
